fix: Correct typo in acerca-de.component.html
Fix a typo in the acerca-de.component.html file where "Desarolladora" was misspelled as "Desarrolladora". This commit corrects the spelling to ensure accurate information is displayed on the About page.

Recent repository commits show a consistent use of the "fix" prefix for bug fixes and a clear and concise description of the changes made.